Reference

Contents

Reference

Contents

Index

Index

StochasticTailAssignment.AircraftRoutingBase.AbstractActivity

StochasticTailAssignment.AircraftRoutingBase.AbstractLeg

StochasticTailAssignment.AircraftRoutingBase.AbstractRoute

StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule

StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule

StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule

StochasticTailAssignment.AircraftRoutingBase.ForcedChaining

StochasticTailAssignment.AircraftRoutingBase.Immat

StochasticTailAssignment.AircraftRoutingBase.Leg

StochasticTailAssignment.AircraftRoutingBase.Maintenance

StochasticTailAssignment.AircraftRoutingBase.MaskedSchedule

StochasticTailAssignment.AircraftRoutingBase.PartialRoute

StochasticTailAssignment.AircraftRoutingBase.RouteWithMaintenance

Dates.dayofweek

Dates.month

StochasticTailAssignment.AircraftRoutingBase.activity_cost

StochasticTailAssignment.AircraftRoutingBase.activity_indices

StochasticTailAssignment.AircraftRoutingBase.aircraft_type

StochasticTailAssignment.AircraftRoutingBase.aircraft_type

StochasticTailAssignment.AircraftRoutingBase.arrival_airport

StochasticTailAssignment.AircraftRoutingBase.arrival_airport_size

StochasticTailAssignment.AircraftRoutingBase.arrival_hour

StochasticTailAssignment.AircraftRoutingBase.arrival_time

StochasticTailAssignment.AircraftRoutingBase.build_immat_graphs

StochasticTailAssignment.AircraftRoutingBase.build_model

StochasticTailAssignment.AircraftRoutingBase.chaining_cost

StochasticTailAssignment.AircraftRoutingBase.chaining_cost

StochasticTailAssignment.AircraftRoutingBase.chaining_cost

StochasticTailAssignment.AircraftRoutingBase.column_generation

StochasticTailAssignment.AircraftRoutingBase.column_heuristic

StochasticTailAssignment.AircraftRoutingBase.compute_arc_index

StochasticTailAssignment.AircraftRoutingBase.compute_arc_index

StochasticTailAssignment.AircraftRoutingBase.compute_edge_mask

StochasticTailAssignment.AircraftRoutingBase.compute_shortest_route

StochasticTailAssignment.AircraftRoutingBase.compute_shortest_subpath

StochasticTailAssignment.AircraftRoutingBase.decode_routes_from_solution

StochasticTailAssignment.AircraftRoutingBase.decode_solution_from_routes

StochasticTailAssignment.AircraftRoutingBase.departure_airport

StochasticTailAssignment.AircraftRoutingBase.departure_airport_size

StochasticTailAssignment.AircraftRoutingBase.departure_hour

StochasticTailAssignment.AircraftRoutingBase.departure_time

StochasticTailAssignment.AircraftRoutingBase.duration

StochasticTailAssignment.AircraftRoutingBase.edge_cost

StochasticTailAssignment.AircraftRoutingBase.edge_cost_origin

StochasticTailAssignment.AircraftRoutingBase.end_airport

StochasticTailAssignment.AircraftRoutingBase.end_airport

StochasticTailAssignment.AircraftRoutingBase.end_hour

StochasticTailAssignment.AircraftRoutingBase.end_time

StochasticTailAssignment.AircraftRoutingBase.end_time

StochasticTailAssignment.AircraftRoutingBase.flight_distance

StochasticTailAssignment.AircraftRoutingBase.flight_distance

StochasticTailAssignment.AircraftRoutingBase.get_activity

StochasticTailAssignment.AircraftRoutingBase.get_last_activity_index

StochasticTailAssignment.AircraftRoutingBase.get_maintenance_indices

StochasticTailAssignment.AircraftRoutingBase.get_s

StochasticTailAssignment.AircraftRoutingBase.get_s

StochasticTailAssignment.AircraftRoutingBase.get_t

StochasticTailAssignment.AircraftRoutingBase.get_t

StochasticTailAssignment.AircraftRoutingBase.highs_model

StochasticTailAssignment.AircraftRoutingBase.id

StochasticTailAssignment.AircraftRoutingBase.immat

StochasticTailAssignment.AircraftRoutingBase.immat_indices

StochasticTailAssignment.AircraftRoutingBase.is_dummy_vertex

StochasticTailAssignment.AircraftRoutingBase.is_feasible

StochasticTailAssignment.AircraftRoutingBase.is_feasible

StochasticTailAssignment.AircraftRoutingBase.is_intersection

StochasticTailAssignment.AircraftRoutingBase.is_leg

StochasticTailAssignment.AircraftRoutingBase.is_location_compatible

StochasticTailAssignment.AircraftRoutingBase.is_maintenance

StochasticTailAssignment.AircraftRoutingBase.is_maintenance

StochasticTailAssignment.AircraftRoutingBase.is_maintenance

StochasticTailAssignment.AircraftRoutingBase.is_s

StochasticTailAssignment.AircraftRoutingBase.is_t

StochasticTailAssignment.AircraftRoutingBase.is_time_compatible

StochasticTailAssignment.AircraftRoutingBase.is_time_compatible

StochasticTailAssignment.AircraftRoutingBase.is_valid_chaining

StochasticTailAssignment.AircraftRoutingBase.leg_indices

StochasticTailAssignment.AircraftRoutingBase.leg_indices_from_route

StochasticTailAssignment.AircraftRoutingBase.mask_graph!

StochasticTailAssignment.AircraftRoutingBase.minimum_turn_time

StochasticTailAssignment.AircraftRoutingBase.minimum_turn_time

StochasticTailAssignment.AircraftRoutingBase.nb_activities

StochasticTailAssignment.AircraftRoutingBase.nb_immats

StochasticTailAssignment.AircraftRoutingBase.nb_legs

StochasticTailAssignment.AircraftRoutingBase.nb_maintenances

StochasticTailAssignment.AircraftRoutingBase.onehot_aircraft_type

StochasticTailAssignment.AircraftRoutingBase.onehot_arrival_airport

StochasticTailAssignment.AircraftRoutingBase.onehot_arrival_hour

StochasticTailAssignment.AircraftRoutingBase.onehot_dayofweek

StochasticTailAssignment.AircraftRoutingBase.onehot_departure_airport

StochasticTailAssignment.AircraftRoutingBase.onehot_departure_hour

StochasticTailAssignment.AircraftRoutingBase.onehot_end_airport

StochasticTailAssignment.AircraftRoutingBase.onehot_end_hour

StochasticTailAssignment.AircraftRoutingBase.onehot_month

StochasticTailAssignment.AircraftRoutingBase.onehot_start_airport

StochasticTailAssignment.AircraftRoutingBase.onehot_start_hour

StochasticTailAssignment.AircraftRoutingBase.operational_cost

StochasticTailAssignment.AircraftRoutingBase.read_instance

StochasticTailAssignment.AircraftRoutingBase.route_to_partial_routes

StochasticTailAssignment.AircraftRoutingBase.scip_model

StochasticTailAssignment.AircraftRoutingBase.slack

StochasticTailAssignment.AircraftRoutingBase.slack_with_turn_time

StochasticTailAssignment.AircraftRoutingBase.solve_aircraft_routing

StochasticTailAssignment.AircraftRoutingBase.start_airport

StochasticTailAssignment.AircraftRoutingBase.start_airport

StochasticTailAssignment.AircraftRoutingBase.start_hour

StochasticTailAssignment.AircraftRoutingBase.start_time

StochasticTailAssignment.AircraftRoutingBase.start_time

StochasticTailAssignment.AircraftRoutingBase.write_instance

StochasticTailAssignment.AircraftRoutingBase.AbstractActivity — Type

abstract type AbstractActivity

Abstract type defining an activity.

Two implementations

Leg

Maintenance

Common methods

is_maintenance

start_airport

end_airport

start_time

end_time

StochasticTailAssignment.AircraftRoutingBase.AbstractLeg — Type

abstract type AbstractLeg <: StochasticTailAssignment.AircraftRoutingBase.AbstractActivity

Abstract type defining a leg

StochasticTailAssignment.AircraftRoutingBase.AbstractRoute — Type

abstract type AbstractRoute

Abstract type for a route.

StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule — Type

abstract type AbstractSchedule

Data structure defining a schedule of activities. Two implementations :

ActivitySchedule : the main data structure, containing all the information about the schedule.

MaskedSchedule : a masked version of the ActivitySchedule, where some activities are removed.

StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule — Type

struct ActivitySchedule{G<:MetaGraphsNext.MetaGraph, G1<:MetaGraphsNext.MetaGraph, L<:StochasticTailAssignment.AircraftRoutingBase.AbstractLeg, A} <: StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule

Data structure defining a schedule of activities. It corresponds to an instance of the aircraft routing (fleet assignment) problem.

Fields

immats::Vector{StochasticTailAssignment.AircraftRoutingBase.Immat}: all available immatriculations

legs::Vector{L} where L<:StochasticTailAssignment.AircraftRoutingBase.AbstractLeg: all legs that need to be operated in the schedule

maintenances::Vector{StochasticTailAssignment.AircraftRoutingBase.Maintenance}: all maintenance task that need to be performed in the schedule

forced_chainings::Vector{StochasticTailAssignment.AircraftRoutingBase.ForcedChaining}: all forced chaining enforced in the schedule

graph::MetaGraphsNext.MetaGraph: dependency (acyclic) digraph between activities

immat_graphs::Vector{G1} where G1<:MetaGraphsNext.MetaGraph: specialized digraph for each immatriculation, removing useless edges

TTM_factor::Float64: minimum turn time multiplier that was used to compute graph edges

include_chaining_costs::Bool: whether to include chaining costs in the operational costs

arc_index::Any: [u, v, i] = arc index for learning

nb_interior_arcs::Int64: number of (u, v, i) not connected to s or t

StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule — Method

ActivitySchedule(
;
    immats,
    legs,
    maintenances,
    forced_chainings,
    fuel_cost_per_kg,
    standard_consumption,
    tractage_cost,
    empty_graph,
    TTM_factor,
    include_chaining_costs,
    store_arc_index,
    sparsification_threshold
)

Constructor for an ActivitySchedule.

StochasticTailAssignment.AircraftRoutingBase.ForcedChaining — Type

struct ForcedChaining

Data structure defining an forced chaining between two activities.

Fields

first_activity_id::String: id of the first activity

second_activity_id::String: id of the second activity chained to the first one

StochasticTailAssignment.AircraftRoutingBase.Immat — Type

struct Immat

Data structure defining an aircraft immatriculation.

Fields

id::String: unique identifier string

fuel_factor::Float64: fuel factor, the percentage of extra fuel consumption for the aircraft type

last_activity_id::String: id of the last activity operated by this immatriculation

aircraft_type::String: aircraft type code of the immatriculation

StochasticTailAssignment.AircraftRoutingBase.Leg — Type

struct Leg <: StochasticTailAssignment.AircraftRoutingBase.AbstractLeg

Fields

id::String: Unique identifier

departure_airport::String: IATA code of the departure airport

arrival_airport::String: IATA code of the arrival airport

departure_time::Dates.DateTime: departure time in UTC

arrival_time::Dates.DateTime: arrival time in UTC

aircraft_type::String: aircraft type code

minimum_turn_time::Int64: minimum rurn time in minutes

StochasticTailAssignment.AircraftRoutingBase.Maintenance — Type

struct Maintenance <: StochasticTailAssignment.AircraftRoutingBase.AbstractActivity

Fields

id::String: unique identifier

location::String: IATA code of the airport

start_time::Dates.DateTime: start time of the maintenance in UTC

end_time::Dates.DateTime: end time of the maintenance in UTC

immat::String: immat code of the aircraft conserned by the maintenance

minimum_turn_time::Int64: minimum turn time in minutes

StochasticTailAssignment.AircraftRoutingBase.MaskedSchedule — Type

struct MaskedSchedule{T<:StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule, G<:MetaGraphsNext.MetaGraph, G1<:MetaGraphsNext.MetaGraph} <: StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule

Wrapper over ActivitySchedule to represent a schedule where some routes are fixed. For each route, the associared immat does not need to be considered anymore, as well as the activities in the route.

Fields

schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule: associated schedule

immat_mask::BitVector: mask of immats that are not fixed

activity_mask::BitVector: mask of activities that are not fixed

leg_mask::Any: mask of legs that are not fixed

graph::MetaGraphsNext.MetaGraph: connection graph

immat_graphs::Vector{G1} where G1<:MetaGraphsNext.MetaGraph: connection sub graphs

StochasticTailAssignment.AircraftRoutingBase.PartialRoute — Type

struct PartialRoute

This data structure enables defining a partial route with a time window. This is used to split a route between maintenances, to only consider legs.

Fields

leg_route::StochasticTailAssignment.AircraftRoutingBase.Route: list of leg indices corresponding to this partial route

start_time::Dates.DateTime: start of the time window for this partial route

end_time::Dates.DateTime: end of the time window for this partial route

StochasticTailAssignment.AircraftRoutingBase.RouteWithMaintenance — Type

struct RouteWithMaintenance <: StochasticTailAssignment.AircraftRoutingBase.AbstractRoute

Route data structure with useful maintenance information.

Fields

route::StochasticTailAssignment.AircraftRoutingBase.Route: a route containing only leg indices

maintenances::Vector{Vector{Int64}}: for each leg in route, ordered list of maintenance indices before the next leg

Dates.dayofweek — Method

dayofweek(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Int64

Retrieve the day of week of activity start time.

Dates.month — Method

month(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Int64

Retrieve the month of activity start time.

StochasticTailAssignment.AircraftRoutingBase.activity_cost — Method

activity_cost(
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule,
    v,
    i
) -> Any

Retrieve the cost of operating activity (usually a leg) corresponding to index v with aircraft immatriculation (index) i in the graph.

StochasticTailAssignment.AircraftRoutingBase.activity_indices — Method

activity_indices(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> UnitRange{Int64}

Retrieve all activity indices in the schedule. Useful for iteration.

StochasticTailAssignment.AircraftRoutingBase.aircraft_type — Method

aircraft_type(
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> String

Returns the aircraft type of the first leg in the schedule, assuming all legs have the same aircraft type.

StochasticTailAssignment.AircraftRoutingBase.aircraft_type — Method

aircraft_type(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> String

Retrieve the aircraft type of given leg.

StochasticTailAssignment.AircraftRoutingBase.arrival_airport — Method

arrival_airport(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> String

Retrieve the arrival airport of given leg.

StochasticTailAssignment.AircraftRoutingBase.arrival_airport_size — Method

arrival_airport_size(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    airport_size_dict::Dict
) -> Any

Retrieve the arrival airport size for given activity.

StochasticTailAssignment.AircraftRoutingBase.arrival_hour — Function

Alias for end_hour.

StochasticTailAssignment.AircraftRoutingBase.arrival_time — Method

arrival_time(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> Dates.DateTime

Retrieve the arrival time of given leg.

StochasticTailAssignment.AircraftRoutingBase.build_immat_graphs — Method

build_immat_graphs(
    graph::MetaGraphsNext.MetaGraph,
    immats::Vector{StochasticTailAssignment.AircraftRoutingBase.Immat},
    maintenances::Vector{StochasticTailAssignment.AircraftRoutingBase.Maintenance},
    legs::Vector{<:StochasticTailAssignment.AircraftRoutingBase.AbstractLeg};
    empty_graph
) -> Vector

Build a subgraph for each immatriculation, corresponding to the routing graph for that immatriculation. Use the same base graph for all immat graphs, in order to have the same vertex indices.

StochasticTailAssignment.AircraftRoutingBase.build_model — Method

build_model(
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule;
    model_builder,
    include_chaining_costs,
    silent,
    feasibility_only,
    relaxation
) -> JuMP.Model

Build the MIP model.

StochasticTailAssignment.AircraftRoutingBase.chaining_cost — Method

chaining_cost(
    ::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    ::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity;
    tractage_cost
) -> Float64

Return the chaining cost between two successive activities.

StochasticTailAssignment.AircraftRoutingBase.chaining_cost — Method

chaining_cost(
    previous_leg::StochasticTailAssignment.AircraftRoutingBase.AbstractLeg,
    next_leg::StochasticTailAssignment.AircraftRoutingBase.AbstractLeg;
    tractage_cost
) -> Float64

Return the chaining cost between two successive legs.

StochasticTailAssignment.AircraftRoutingBase.chaining_cost — Method

chaining_cost(
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule,
    u,
    v
) -> Any

Retrieve (from the graph) the cost of chaining from vertex u to vertex v.

StochasticTailAssignment.AircraftRoutingBase.column_generation — Method

column_generation(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    initial_paths::Vector{StochasticTailAssignment.AircraftRoutingBase.Route};
    model_builder,
    include_chaining_costs,
    max_iterations,
    tol,
    silent
) -> Tuple{Vector{StochasticTailAssignment.AircraftRoutingBase.Route}, JuMP.Model, Vector{Vector{Float64}}}

Perform column generation to solve the deterministic tail assignment problem.

Returns:

columns: vector of generated columns as a Vector{Route}

model: final JuMP model

VV: reduced cost evolution for each aircraft

StochasticTailAssignment.AircraftRoutingBase.column_heuristic — Method

column_heuristic(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    columns::Vector{StochasticTailAssignment.AircraftRoutingBase.Route};
    model_builder,
    include_chaining_costs,
    silent,
    bin,
    column_cost,
    time_limit,
    warm_start
) -> Tuple{Vector{StochasticTailAssignment.AircraftRoutingBase.Route}, Union{Float64, Vector{Float64}}}

Take a vector of Routes and returns the solution of the restricted master problem heuristic for the deterministic tail assignment problem.

StochasticTailAssignment.AircraftRoutingBase.compute_arc_index — Method

compute_arc_index(
    immat_graphs,
    graph
) -> Tuple{Dict{Tuple{Int64, Int64, Int64}, Int64}, Int64}

Compute the arc index for the given graph and schedule. [u, v, i] -> arc index

StochasticTailAssignment.AircraftRoutingBase.compute_arc_index — Method

compute_arc_index(
    schedule::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule
) -> Tuple{Dict{Tuple{Int64, Int64, Int64}, Int64}, Int64}

Compute the arc index for the given schedule.

StochasticTailAssignment.AircraftRoutingBase.compute_edge_mask — Method

compute_edge_mask(
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> Any

Returns a boolean mask of the edges that connect two activity (i.e. neither of the dummy vertices).

StochasticTailAssignment.AircraftRoutingBase.compute_shortest_route — Method

compute_shortest_route(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    immat_index::Int64,
    λ_val;
    include_chaining_costs,
    silent
)

Solve the pricing problem for given aircraft immat_index and dual variables λ_val. This is solved as successive shortest paths between scheduled maintenances for immat_index.

StochasticTailAssignment.AircraftRoutingBase.compute_shortest_subpath — Method

compute_shortest_subpath(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    start_node,
    end_node,
    weights
) -> Vector

Compute a shortest path between start_node and end_node, with weight matrix weights. Uses the ford bellman algorithm, since the graph is acyclic.

StochasticTailAssignment.AircraftRoutingBase.decode_routes_from_solution — Method

decode_routes_from_solution(
    y_val,
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule
) -> Vector{StochasticTailAssignment.AircraftRoutingBase.Route}

Decode routes from the binary variables of the solution of the MIP model.

StochasticTailAssignment.AircraftRoutingBase.decode_solution_from_routes — Method

decode_solution_from_routes(
    routes,
    instance
) -> Union{Array{Float64, 3}, Matrix}

Recompute binary variables solution of the MIP model from given routes.

StochasticTailAssignment.AircraftRoutingBase.departure_airport — Method

departure_airport(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> String

Retrieve the departure airport of given leg.

StochasticTailAssignment.AircraftRoutingBase.departure_airport_size — Method

departure_airport_size(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    airport_size_dict::Dict
) -> Any

Retrieve the departure airport size for given activity.

StochasticTailAssignment.AircraftRoutingBase.departure_hour — Function

Alias for start_hour.

StochasticTailAssignment.AircraftRoutingBase.departure_time — Method

departure_time(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> Dates.DateTime

Retrieve the departure time of given leg.

StochasticTailAssignment.AircraftRoutingBase.duration — Method

duration(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Float64

Compute the duration of an activity in minutes.

StochasticTailAssignment.AircraftRoutingBase.edge_cost — Method

edge_cost(
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule,
    u::Int64,
    v::Int64,
    i::Int64;
    include_chaining_costs
) -> Any

Compute the cost of an edge as the sum of the operational cost of the destination plus the eventual chaining cost of the edge.

StochasticTailAssignment.AircraftRoutingBase.edge_cost_origin — Method

edge_cost_origin(
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule,
    u::Int64,
    v::Int64,
    i::Int64;
    include_chaining_costs
) -> Any

Compute the cost of an edge as the sum of the operational cost of the origin plus the chaining cost of the edge.

StochasticTailAssignment.AircraftRoutingBase.end_airport — Method

end_airport(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> String

Retrieve the end airport of given leg.

StochasticTailAssignment.AircraftRoutingBase.end_airport — Method

end_airport(
    maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance
) -> String

Retrieve the location of given maintenance.

StochasticTailAssignment.AircraftRoutingBase.end_hour — Method

end_hour(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Int64

Retrieve the hour of activity start time.

StochasticTailAssignment.AircraftRoutingBase.end_time — Method

end_time(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> Dates.DateTime

Retrieve the arrival time of given leg.

StochasticTailAssignment.AircraftRoutingBase.end_time — Method

end_time(
    maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance
) -> Dates.DateTime

Retrieve the end time of given maintenance.

StochasticTailAssignment.AircraftRoutingBase.flight_distance — Method

flight_distance(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg,
    distance_dict::Dict
) -> Any

Retrieve the flight distance for given leg.

StochasticTailAssignment.AircraftRoutingBase.flight_distance — Method

flight_distance(
    _::StochasticTailAssignment.AircraftRoutingBase.Maintenance,
    _::Dict
) -> Float64

Retrieve the flight distance for given leg.

StochasticTailAssignment.AircraftRoutingBase.get_activity — Method

get_activity(
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    activity_index::Int64
) -> Union{StochasticTailAssignment.AircraftRoutingBase.AbstractLeg, StochasticTailAssignment.AircraftRoutingBase.Maintenance}

Retrieve the activity corresponding to the given index.

Warning

This method might be type instable. It can return either a Leg or a Maintenance, depending on activity_index value.

StochasticTailAssignment.AircraftRoutingBase.get_last_activity_index — Method

get_last_activity_index(
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    immat_index::Int64
) -> Integer

Return last activity index of immat if it exists, else index of s.

StochasticTailAssignment.AircraftRoutingBase.get_maintenance_indices — Method

get_maintenance_indices(
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    immat_index::Int64;
    sorted
) -> Vector{Int64}

Return a vector of node indices corresponding to all maintenances affected to aircraft immat_index. Sorted by start time by default.

StochasticTailAssignment.AircraftRoutingBase.get_s — Method

get_s(graph::MetaGraphsNext.MetaGraph) -> Integer

Retrieve the vertex code of the source vertex s in the graph.

StochasticTailAssignment.AircraftRoutingBase.get_s — Method

get_s(
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule
) -> Integer

Retrieve the vertex code of the source vertex s in the graph.

StochasticTailAssignment.AircraftRoutingBase.get_t — Method

get_t(graph::MetaGraphsNext.MetaGraph) -> Integer

Retrieve the vertex code of the sink vertex t in the graph.

StochasticTailAssignment.AircraftRoutingBase.get_t — Method

get_t(
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule
) -> Integer

Retrieve the vertex code of the sink vertex t in the graph.

StochasticTailAssignment.AircraftRoutingBase.highs_model — Method

highs_model() -> JuMP.Model

Initialize a HiGHS model.

StochasticTailAssignment.AircraftRoutingBase.id — Method

id(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Any

Retrieve the id of given activity.

StochasticTailAssignment.AircraftRoutingBase.immat — Method

immat(
    maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance
) -> String

Retrieve the immat code associated to given maintenance.

StochasticTailAssignment.AircraftRoutingBase.immat_indices — Method

immat_indices(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> UnitRange{Int64}

Retrieve all immatriculation indices in the schedule. Useful for iteration.

StochasticTailAssignment.AircraftRoutingBase.is_dummy_vertex — Method

is_dummy_vertex(
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    activity_index::Int64
) -> Any

Check if activity_index is either one of the two dummy vertices of the graph.

StochasticTailAssignment.AircraftRoutingBase.is_feasible — Method

is_feasible(
    route::StochasticTailAssignment.AircraftRoutingBase.Route,
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule;
    verbose
) -> Bool

Check if a given route is feasible.

StochasticTailAssignment.AircraftRoutingBase.is_feasible — Method

is_feasible(
    routes::Vector{StochasticTailAssignment.AircraftRoutingBase.Route},
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule;
    verbose
) -> Bool

Check if routes are feasible.

StochasticTailAssignment.AircraftRoutingBase.is_intersection — Method

is_intersection(
    p1::StochasticTailAssignment.AircraftRoutingBase.PartialRoute,
    p2::StochasticTailAssignment.AircraftRoutingBase.PartialRoute
) -> Bool

Check if two partial routes intersect in time, i.e. if they are elegible for a swap.

StochasticTailAssignment.AircraftRoutingBase.is_leg — Method

is_leg(
    schedule::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule,
    activity_index::Int64
) -> Any

Check if given activity index activity_index corresponds to a leg activity.

StochasticTailAssignment.AircraftRoutingBase.is_location_compatible — Method

is_location_compatible(
    previous_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    next_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Bool

Check if two successive activities are location compatible, i.e. if the arrival airport of previous_activity is the departure airport of next_activity.

StochasticTailAssignment.AircraftRoutingBase.is_maintenance — Method

is_maintenance(
    schedule::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule,
    activity_index::Int64;
    safe_mode
) -> Any

Check if given activity index activity_index corresponds to a maintenance.

StochasticTailAssignment.AircraftRoutingBase.is_maintenance — Method

is_maintenance(
    _::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> Bool

Check if an activity is a maintenance.

StochasticTailAssignment.AircraftRoutingBase.is_maintenance — Method

is_maintenance(
    maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance
) -> Bool

Check if an activity is a maintenance.

StochasticTailAssignment.AircraftRoutingBase.is_s — Method

is_s(schedule, activity_index::Int64) -> Any

Check if activity_index is the source of the graph.

StochasticTailAssignment.AircraftRoutingBase.is_t — Method

is_t(schedule, activity_index::Int64) -> Any

Check if activity_index is the sink of the graph.

StochasticTailAssignment.AircraftRoutingBase.is_time_compatible — Method

is_time_compatible(
    previous_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    next_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity;
    TTM_factor
) -> Bool

Check if two successive activities are time compatible, i.e. if the slack between them is larger than the minimumturntime of next_activity.

StochasticTailAssignment.AircraftRoutingBase.is_time_compatible — Method

is_time_compatible(
    previous_maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance,
    next_maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance;
    TTM_factor
) -> Bool

Check if two successive maintenances are time compatible.

StochasticTailAssignment.AircraftRoutingBase.is_valid_chaining — Method

is_valid_chaining(
    previous_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    next_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity;
    verbose,
    TTM_factor
) -> Bool

Check if a given activity chaining (two successive activities) are compatible. Additionaly prints a warning if the chaining is not compatible and verbose is true.

StochasticTailAssignment.AircraftRoutingBase.leg_indices — Method

leg_indices(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> UnitRange{Int64}

Retrieve all leg indices in the schedule. Useful for iteration.

StochasticTailAssignment.AircraftRoutingBase.leg_indices_from_route — Method

leg_indices_from_route(
    route,
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule
) -> Vector

Returns an ordered vector of leg activity indices in route (i.e. filtering out maintenances).

StochasticTailAssignment.AircraftRoutingBase.mask_graph! — Method

mask_graph!(
    graph,
    route::StochasticTailAssignment.AircraftRoutingBase.Route;
    verbose
)

Remove edges connected to masked activities (this allows keeping consistent vertex indexing)

StochasticTailAssignment.AircraftRoutingBase.minimum_turn_time — Method

minimum_turn_time(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Float64

Retrieve the minimum turn time for an activity.

StochasticTailAssignment.AircraftRoutingBase.minimum_turn_time — Method

minimum_turn_time(
    maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance
) -> Float64

Retrieve the minimum turn time of given maintenance.

StochasticTailAssignment.AircraftRoutingBase.nb_activities — Method

nb_activities(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> Int64

Retrieve the total number of activities (legs + maintenances) in the schedule.

StochasticTailAssignment.AircraftRoutingBase.nb_immats — Method

nb_immats(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> Int64

Retrieve the number of immatriculations in the schedule.

StochasticTailAssignment.AircraftRoutingBase.nb_legs — Method

nb_legs(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> Int64

Retrieve the number of legs in the schedule.

StochasticTailAssignment.AircraftRoutingBase.nb_maintenances — Method

nb_maintenances(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> Int64

Retrieve the number of maintenances in the schedule.

StochasticTailAssignment.AircraftRoutingBase.onehot_aircraft_type — Method

onehot_aircraft_type(
    leg::StochasticTailAssignment.AircraftRoutingBase.AbstractLeg,
    aircraft_types::AbstractVector
) -> Vector{Bool}

Encode the aircraft type of given leg as a one-hot vector.

aircraft_types should the list of all possible aircraft types. If the activity type is not in the list, it is encoded as the "other" category.

StochasticTailAssignment.AircraftRoutingBase.onehot_arrival_airport — Function

Alias for onehot_end_airport.

StochasticTailAssignment.AircraftRoutingBase.onehot_arrival_hour — Function

Alias for onehot_end_hour.

StochasticTailAssignment.AircraftRoutingBase.onehot_dayofweek — Function

onehot_dayofweek(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Vector{Bool}
onehot_dayofweek(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    range
) -> Vector{Bool}

Encode the day of week of given activity as a one-hot vector.

StochasticTailAssignment.AircraftRoutingBase.onehot_departure_airport — Function

Alias for onehot_start_airport.

StochasticTailAssignment.AircraftRoutingBase.onehot_departure_hour — Function

Alias for onehot_start_hour.

StochasticTailAssignment.AircraftRoutingBase.onehot_end_airport — Method

onehot_end_airport(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    airports::AbstractVector
) -> OneHotArrays.OneHotVector{UInt32}

Encode the end airport of given activity as a one-hot vector.

airports should the list of all possible airports. If the activity airport is not in the list, it is encoded as the "other" category.

StochasticTailAssignment.AircraftRoutingBase.onehot_end_hour — Function

onehot_end_hour(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Vector{Bool}
onehot_end_hour(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    range
) -> Vector{Bool}

Encode the and hour of given activity as a one-hot vector.

StochasticTailAssignment.AircraftRoutingBase.onehot_month — Function

onehot_month(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Vector{Bool}
onehot_month(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    range
) -> Vector{Bool}

Encode the start month of given activity as a one-hot vector.

StochasticTailAssignment.AircraftRoutingBase.onehot_start_airport — Method

onehot_start_airport(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    airports::AbstractVector
) -> OneHotArrays.OneHotVector{UInt32}

Encode the start airport of given activity as a one-hot vector.

airports should the list of all possible airports. If the activity airport is not in the list, it is encoded as the "other" category.

StochasticTailAssignment.AircraftRoutingBase.onehot_start_hour — Function

onehot_start_hour(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Vector{Bool}
onehot_start_hour(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    range
) -> Vector{Bool}

Encode the start hour of given activity as a one-hot vector.

StochasticTailAssignment.AircraftRoutingBase.operational_cost — Method

operational_cost(
    routes::Vector{Vector{Int64}},
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule;
    include_chaining_costs
) -> Any

Compute the total cost of operating all activities in the routes routes.

StochasticTailAssignment.AircraftRoutingBase.read_instance — Method

read_instance(
    path;
    standard_consumption,
    TTM_factor,
    include_chaining_costs,
    kwargs...
)

Read instance from custom json data format at path.

StochasticTailAssignment.AircraftRoutingBase.route_to_partial_routes — Method

route_to_partial_routes(
    route::StochasticTailAssignment.AircraftRoutingBase.Route,
    schedule::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
) -> Vector{StochasticTailAssignment.AircraftRoutingBase.PartialRoute}

Convert a given feasible route into a vector of partial routes splitting the maintenances.

StochasticTailAssignment.AircraftRoutingBase.scip_model — Method

scip_model() -> JuMP.Model

Initialize a SCIP model.

StochasticTailAssignment.AircraftRoutingBase.slack — Method

slack(
    previous_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    next_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Float64

Return the slack time between two successive activities.

Warning

This does not include the minimum turn time. For that, use slack_with_turn_time.

StochasticTailAssignment.AircraftRoutingBase.slack_with_turn_time — Method

slack_with_turn_time(
    previous_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity,
    next_activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Float64

Return the slack time between two successive activities, including the minimum turn time of the second activity. slack(l₁, l₂) = (scheduled duration between arrival of l1 and departure of l2) - (minimum turn time)

StochasticTailAssignment.AircraftRoutingBase.solve_aircraft_routing — Method

solve_aircraft_routing(
    instance::StochasticTailAssignment.AircraftRoutingBase.AbstractSchedule;
    relaxation,
    silent,
    kwargs...
) -> Tuple{Union{Nothing, Vector{StochasticTailAssignment.AircraftRoutingBase.Route}, Vector{Vector{Int64}}}, Union{Float64, Int64, Vector{Float64}}, Any}

Solve the aircraft routing problem using a MIP solver. Returns the optimal routes and objective value.

StochasticTailAssignment.AircraftRoutingBase.start_airport — Method

start_airport(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> String

Retrieve the start airport of given leg.

StochasticTailAssignment.AircraftRoutingBase.start_airport — Method

start_airport(
    maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance
) -> String

Retrieve the location of given maintenance.

StochasticTailAssignment.AircraftRoutingBase.start_hour — Method

start_hour(
    activity::StochasticTailAssignment.AircraftRoutingBase.AbstractActivity
) -> Int64

Retrieve the hour of activity start time.

StochasticTailAssignment.AircraftRoutingBase.start_time — Method

start_time(
    leg::StochasticTailAssignment.AircraftRoutingBase.Leg
) -> Dates.DateTime

Retrieve the departure time of given leg.

StochasticTailAssignment.AircraftRoutingBase.start_time — Method

start_time(
    maintenance::StochasticTailAssignment.AircraftRoutingBase.Maintenance
) -> Dates.DateTime

Retrieve the start time of given maintenance.

StochasticTailAssignment.AircraftRoutingBase.write_instance — Function

write_instance(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule
)
write_instance(
    instance::StochasticTailAssignment.AircraftRoutingBase.ActivitySchedule,
    path
)

Write given input schedule to custom json data format at path.